api/pyproject.toml
[tool.poetry]
name = "admin-dashboard-api"
version = "0.1.0"
description = "Backend API for Admin Dashboard"
authors = ["Lillith Long"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pydantic = "^2.6.1"
odmantic = "^1.0.2"
motor = "^3.3.2"
pymongo = "^4.6.1"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
bcrypt = "^4.0.1"
python-dotenv = "^1.0.1"
email-validator = "^2.0.0"
httpx = "^0.26.0"
loguru = "^0.7.2"
tenacity = "^8.2.3"
fastapi-pagination = "^0.12.14"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
black = "^24.1.1"
isort = "^5.13.2"
ruff = "^0.2.1"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
faker = "^22.6.0"
factory-boy = "^3.3.0"
httpx = "^0.26.0"
asgi-lifespan = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.8.1"]
build-backend = "poetry.core.masonry.api"